% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Generate Multivariate Normal Data
\eqn{\mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma}\right)}}
\usage{
mvn(n, mu = NULL, Sigma, R = NULL, par = FALSE, ncores = NULL, ...)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{mu}{Numeric vector.
Location parameter mean vector of length \eqn{k}
(\eqn{\boldsymbol{\mu}}).}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}

\item{R}{Integer.
Number of Monte Carlo replications.
By default,
\code{R = NULL} produces a single random data set.
If \code{R} is defined
(e.g., \code{R = 10}),
produces a list of length \code{R}
of multiple random data sets.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}

\item{...}{Arguments that can be passed to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\value{
If \code{R = NULL},
returns an \eqn{n \times k} multivariate normal data matrix generated
using the mean vector
\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
and the variance-covariance matrix
\code{Sigma}
\eqn{\left( \boldsymbol{\Sigma} \right)}
provided.
If \code{R} is an integer,
returns a list of length \code{R} of
\eqn{n \times k} multivariate normal data matrices generated
using the mean vector
\verb{mu`` and the variance-covariance matrix }Sigma`
provided.
}
\description{
Generates an \eqn{n \times k}
multivariate normal data
matrix
\eqn{\mathbf{X}}
from a
\eqn{k \times 1} mean vector
\eqn{\boldsymbol{\mu}}
and a
\eqn{k \times k} variance-covariance matrix
\eqn{\boldsymbol{\Sigma}}.
This function is a wrapper around
\code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.
}
\details{
The multivariate normal
(or multivariate Gaussian,
or joint normal)
distribution is given by
\deqn{
  f_{\mathbf{X}}
  \left(
    X_1,
    \dots,
    X_k
  \right)
  =
  \frac{
    \exp
    \left[
      -
      \frac{1}{2}
      \left(
        \mathbf{X} - \boldsymbol{\mu}
      \right)^{\prime}
      \boldsymbol{\Sigma}^{-1}
      \left(
        \mathbf{X} - \boldsymbol{\mu}
      \right)
    \right]
  }{
    \sqrt{
      \left(
        2 \pi
      \right)^{k}
      | \boldsymbol{\Sigma} |
    }
  }
}
with
\eqn{k}-dimensional
random vector
\eqn{\mathbf{X} \in \boldsymbol{\mu} + \textrm{span}\left(\boldsymbol{\Sigma}\right) \subseteq \mathbf{R}^k},
\eqn{\boldsymbol{\mu}}
is the location parameter mean
(\eqn{\boldsymbol{\mu} \in \mathbf{R}^k}),
and
\eqn{\boldsymbol{\Sigma}}
is the variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \in \mathbf{R}^{k \times k}}).
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
# single random data set
X <- mvn(
  n = 100,
  mu = mu,
  Sigma = Sigma
)
str(X, list.len = 6)
# multiple random data sets
X <- mvn(
  n = 100,
  mu = mu,
  Sigma = Sigma,
  R = 10
)
str(X, list.len = 6)
}
\references{
Venables, W. N., Ripley, B. D., & Venables, W. N. (2002).
\emph{Modern applied statistics with S}.
New York, N.Y: Springer.

\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate normal distribution}
}
\seealso{
\code{\link[jeksterslabRdist:mvnpdf]{jeksterslabRdist::mvnpdf()}},
\code{\link[jeksterslabRdist:mvnll]{jeksterslabRdist::mvnll()}},
and
\code{\link[jeksterslabRdist:mvn2ll]{jeksterslabRdist::mvn2ll()}},
for more information on the multivariate normal distribution.

Other multivariate data functions: 
\code{\link{mvnram}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate data functions}
\keyword{multivariate}
\keyword{normal}
