% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvnram}
\alias{mvnram}
\title{Generate Multivariate Normal Data from RAM Matrices}
\usage{
mvnram(
  n,
  A,
  S,
  F,
  I,
  M = NULL,
  mu = NULL,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  ...
)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{A}{Matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{Matrix.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{F}{Matrix.
Filter matrix
used to select the observed variables.}

\item{I}{Matrix.
Identity matrix.}

\item{M}{Numeric vector.
Mean structure.
Vector of means and intercepts.}

\item{mu}{Numeric vector.
Location parameter mean vector of length \eqn{k}
(\eqn{\boldsymbol{\mu}}).}

\item{R}{Integer.
Number of Monte Carlo replications.
By default,
\code{R = NULL} produces a single random data set.
If \code{R} is defined
(e.g., \code{R = 10}),
produces a list of length \code{R}
of multiple random data sets.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{...}{Arguments that can be passed to \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.}
}
\value{
If \code{R = NULL},
returns an \eqn{n \times k} multivariate normal data matrix generated
using the provided or derived mean vector
\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
and the derived variance-covariance matrix
\code{Sigma}
\eqn{\left( \boldsymbol{\Sigma} \right)}.
If \code{R} is an integer,
returns a list of length \code{R} of
\eqn{n \times k} multivariate normal data matrices generated
using the provided or derived mean vector
\code{mu}
and the derived variance-covariance matrix
\code{Sigma}.
}
\description{
Generates multivariate normal data from
the Reticular Action Model (RAM) matrices.
}
\details{
The multivariate normal distribution has two parameters,
namely the
\eqn{k \times 1} mean vector
\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
and the
\eqn{k \times k} variance-covariance matrix
\code{Sigma}
\eqn{\left( \boldsymbol{\Sigma} \right)}.

The mean vector
\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
can be supplied directly using the
\code{mu} argument.
It can also be derived using
\code{M} .
Note that the argument
\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
takes precedence over
\code{M}
If
\code{mu}
is not provided,
it is computed using
\code{M}
with the
\code{\link[jeksterslabRsem:ram_mutheta]{jeksterslabRsem::ram_mutheta()}}
function.
If both
\code{mu}
and
\code{M}
are not provided,
\code{mu}
is set to a vector of zeroes with the appropriate length.

The variance-covariance matrix \code{Sigma}
\eqn{\left(\boldsymbol{\Sigma}\right)}
is derived
from the RAM matrices \code{A}, \code{S}, \code{F}, and \code{I}.

\code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
and
\code{Sigma}
\eqn{\left(\boldsymbol{\Sigma}\right)}
are then used
by the
\code{\link[=mvn]{mvn()}}
to generate multivariate normal data.
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
# single random data set
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
str(X, list.len = 6)
# multiple random data sets
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu,
  R = 10
)
str(X, list.len = 6)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
\code{\link[jeksterslabRsem:ram_Sigmatheta]{jeksterslabRsem::ram_Sigmatheta()}},
and
\code{\link[jeksterslabRsem:ram_mutheta]{jeksterslabRsem::ram_mutheta()}}
for more information on the Reticular Action Model
and
\code{\link[=mvn]{mvn()}}
for multivariate normal data generation.

Other multivariate data functions: 
\code{\link{mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate data functions}
\keyword{multivariate}
\keyword{normal}
