% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ.R
\name{univ}
\alias{univ}
\title{Generate Univariate Data}
\usage{
univ(n, rFUN = rnorm, R = NULL, par = FALSE, ncores = NULL, ...)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{rFUN}{Function.
Data generating function to generate univariate data.}

\item{R}{Integer.
Number of Monte Carlo replications.
By default,
\code{R = NULL} produces a single random data set.
If \code{R} is defined
(e.g., \code{R = 10}),
produces a list of length \code{R}
of multiple random data sets.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}

\item{...}{Arguments to pass to \code{rFUN}.}
}
\value{
If \code{R = NULL},
returns an \eqn{n \times 1} univariate data vector
generated using \code{rFUN} and parameters passed to \code{...}.
If \code{R} is an integer,
returns a list of length \code{R} of
\eqn{n \times 1} univariate data vector
generated using \code{rFUN} and parameters passed to \code{...}.
}
\description{
Generates an \eqn{n \times 1}
univariate data vector.
}
\details{
The univariate distribution and parameters used in the
data generating process can be specified using \code{rFUN} and \code{...}.
}
\examples{
n <- 5
R <- 5
#################################
# normal distribution
#################################
mu <- 100
sigma2 <- 225
sigma <- sqrt(sigma2)
x <- univ(
  n = n,
  rFUN = rnorm,
  mean = mu,
  sd = sigma
)
str(x)
x <- univ(
  n = n,
  rFUN = rnorm,
  R = R,
  mean = mu,
  sd = sigma
)
str(x)
#################################
# binomial distribution
#################################
n_trials <- 1
p <- 0.50
x <- univ(
  n = n,
  rFUN = rbinom,
  size = n_trials,
  prob = p
)
str(x)
x <- univ(
  n = n,
  rFUN = rbinom,
  R = R,
  size = n_trials,
  prob = p
)
str(x)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{univariate data functions}
\keyword{univariate}
