% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvn}
\alias{mvn}
\title{Generate Multivariate Normal Data
\eqn{
    \mathbf{X} \sim \mathcal{N}_{k}
    \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right)
    %(\#eq:dist-X-mvn)
  }}
\usage{
mvn(
  n,
  mu = NULL,
  Sigma,
  tol = 1e-06,
  empirical = FALSE,
  df = FALSE,
  varnames = NULL,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{mu}{Numeric vector.
Location parameter mean vector \eqn{\boldsymbol{\mu}} of length \eqn{k}.}

\item{Sigma}{Numeric matrix.
\eqn{k \times k} variance-covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{tol}{Numeric.
Tolerance (relative to largest variance)
for numerical lack of positive-definiteness in \code{Sigma}.}

\item{empirical}{Logical.
If \code{TRUE}, \code{mu} and \code{Sigma} specify the empirical,
not population. mean and covariance matrix.}

\item{df}{Logical.
If \code{TRUE}, the function returns a data frame.
If \code{FALSE}, the function returns a matrix.}

\item{varnames}{Character string.
Optional column names with the same length as \code{mu}.}

\item{R}{Integer.
Number of Monte Carlo replications. If \code{R} is not provided,
the function produces a single random data set.
If \code{R} is an integer greater than 1, (e.g., \code{R = 10}),
the function produces multiple random data sets
stored in each element of a list of length \code{R}.
\code{par} and all succeeding arguments are only relevant when \code{R > 1}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\value{
If \code{R = NULL} or \code{R = 1}, returns an \eqn{n \times k}
multivariate normal data matrix or data frame .
If \code{R} is an integer greater than 1, (e.g., \code{R = 10})
returns a list of length \code{R} of \eqn{n \times k}
multivariate normal data matrix or data frame.
}
\description{
Generates an \eqn{n \times k} multivariate data matrix
or a list of \eqn{n \times k} multivariate data matrices of length \code{R}
from the multivariate normal distribution
\deqn{
    \mathbf{X} \sim \mathcal{N}_{k}
    \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) .
    %(\#eq:dist-X-mvn)
  }
This function is a wrapper around \code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.
}
\details{
The multivariate normal distribution has two parameters, namely
the \eqn{k \times 1} mean vector \code{mu} \eqn{\left( \boldsymbol{\mu} \right)}
and the \eqn{k \times k} variance-covariance matrix \code{Sigma}
\eqn{\left( \boldsymbol{\Sigma} \right)}.
If \code{mu} is not provided,
it is set to a vector of zeroes with the appropriate length.

Options for explicit parallelism are provided when \code{R > 1}
especially when \code{R} is large. See \code{par} and suceeding arguments.
}
\examples{
mu <- c(100, 100, 100)
Sigma <- matrix(
  data = c(225, 112.50, 56.25, 112.5, 225, 112.5, 56.25, 112.50, 225),
  ncol = 3
)
X <- mvn(n = 100, mu = mu, Sigma = Sigma)
Xstar <- mvn(n = 100, mu = mu, Sigma = Sigma, R = 100)
}
\references{
Venables, W. N., Ripley, B. D., & Venables, W. N. (2002).
\emph{Modern applied statistics with S}.  New York, N.Y: Springer.

\href{https://en.wikipedia.org/wiki/Multivariate_normal_distribution}{Wikipedia: Multivariate normal distribution}
}
\seealso{
\code{\link[jeksterslabRdist:mvnpdf]{jeksterslabRdist::mvnpdf()}}, \code{\link[jeksterslabRdist:mvnll]{jeksterslabRdist::mvnll()}},
and \code{\link[jeksterslabRdist:mvn2ll]{jeksterslabRdist::mvn2ll()}},
for more information on the multivariate normal distribution.

Other multivariate data functions: 
\code{\link{mvnramsigma2}()},
\code{\link{mvnram}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate data functions}
\keyword{multivariate,}
\keyword{normal}
