% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn.R
\name{mvnram}
\alias{mvnram}
\title{Generate Multivariate Normal Data
\eqn{
  \mathbf{X}
  \sim
  \mathcal{N}_{k}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
  \right)
  %(\#eq:dist-X-mvn)
}
from the Reticular Action Model (RAM) Matrices}
\usage{
mvnram(
  n,
  A,
  S,
  F,
  I,
  M = NULL,
  mu = NULL,
  tol = 1e-06,
  empirical = FALSE,
  R = NULL,
  par = FALSE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars
)
}
\arguments{
\item{n}{Integer.
Sample size.}

\item{A}{\eqn{\mathbf{A}} matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{\eqn{\mathbf{S}} matrix.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{F}{\eqn{\mathbf{F}} matrix.
Filter matrix
used to select the observed variables.}

\item{I}{\eqn{\mathbf{I}} matrix.
Identity matrix.}

\item{M}{Numeric vector.
Mean structure.
Vector of means and intercepts.}

\item{mu}{Numeric vector.
Location parameter mean vector \eqn{\boldsymbol{\mu}} of length \eqn{k}.}

\item{tol}{Numeric.
Tolerance (relative to largest variance)
for numerical lack of positive-definiteness in \code{Sigma}.}

\item{empirical}{Logical.
If \code{TRUE}, \code{mu} and \code{Sigma} specify the empirical
not population mean and covariance matrix.}

\item{R}{Integer.
Number of Monte Carlo replications.
If \code{R} is not provided,
the function produces a single random data set.
If \code{R} is an integer greater than 1,
(e.g., \code{R = 10}),
the function produces
multiple random data sets
stored in each element of a list of length \code{R}.
\code{par} and all succeeding arguments are only relevant
when \code{R > 1}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.
If \code{FALSE}, use \code{\link[=lapply]{lapply()}}.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{mc}{Logical.
If \code{TRUE}, use \code{\link[parallel:mclapply]{parallel::mclapply()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}} or \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
Ignored if \code{par = FALSE}.}

\item{lb}{Logical.
If \code{TRUE} use \code{\link[parallel:clusterApply]{parallel::parLapplyLB()}}.
If \code{FALSE}, use \code{\link[parallel:clusterApply]{parallel::parLapply()}}.
Ignored if \code{par = FALSE} and \code{mc = TRUE}.}

\item{cl_eval}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}} using \code{cl_expr}.
Ignored if \code{mc = TRUE}.}

\item{cl_export}{Logical.
Execute \code{\link[parallel:clusterApply]{parallel::clusterExport()}} using \code{cl_vars}.
Ignored if \code{mc = TRUE}.}

\item{cl_expr}{Expression.
Expression passed to \code{\link[parallel:clusterApply]{parallel::clusterEvalQ()}}
Ignored if \code{mc = TRUE}.}

\item{cl_vars}{Character vector.
Names of objects to pass to \code{\link[parallel:clusterApply]{parallel::clusterExport()}}
Ignored if \code{mc = TRUE}.}
}
\value{
If \code{R = NULL} or \code{R = 1},
returns an \eqn{n \times k} multivariate normal data matrix generated
using the provided or derived mean vector \code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
and the derived variance-covariance matrix \code{Sigma}
\eqn{\left( \boldsymbol{\Sigma} \right)}.
If \code{R} is an integer greater than 1,
(e.g., \code{R = 10})
returns a list of length \code{R} of
\eqn{n \times k} multivariate normal data matrices generated
using the provided or derived mean vector \code{mu}
and the derived variance-covariance matrix \code{Sigma}.
}
\description{
Generates an \eqn{n \times k} multivariate data matrix
or a list of \eqn{n \times k} multivariate data matrices of length \code{R}
from the multivariate normal distribution
\deqn{
  \mathbf{X}
  \sim
  \mathcal{N}_{k}
  \left(
    \boldsymbol{\mu},
    \boldsymbol{\Sigma}
  \right)
  %(\#eq:dist-X-mvn)
}
The model-implied matrices used to generate data
is derived from the Reticular Action Model (RAM) Matrices.
}
\details{
The multivariate normal distribution has two parameters,
namely the
\eqn{k \times 1} mean vector
\code{mu} \eqn{\left( \boldsymbol{\mu} \right)}
and the
\eqn{k \times k} variance-covariance matrix
\code{Sigma} \eqn{\left( \boldsymbol{\Sigma} \right)}.

The mean vector \code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
can be supplied directly
using the \code{mu} argument.
It can also be derived
using \code{M}.
Note that the argument \code{mu}
\eqn{\left( \boldsymbol{\mu} \right)}
takes precedence over \code{M}
If \code{mu} is not provided,
it is computed using \code{M}
with the \code{\link[jeksterslabRsem:ram_mutheta]{jeksterslabRsem::ram_mutheta()}} function.
If both \code{mu} and \code{M}
are not provided,
\code{mu} is set to a vector of zeroes with the appropriate length.

The variance-covariance matrix \code{Sigma}
\eqn{\left(\boldsymbol{\Sigma}\right)}
is derived
from the RAM matrices \code{A}, \code{S}, \code{F}, and \code{I}.

\code{mu} \eqn{\left( \boldsymbol{\mu} \right)}
and \code{Sigma} \eqn{\left(\boldsymbol{\Sigma}\right)}
are then used by the \code{\link[=mvn]{mvn()}}
to generate multivariate normal data.

Options for explicit parallelism are provided
when \code{R > 1} especially when \code{R} is large.
See \code{par} and suceeding arguments.
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
S <- F <- I <- diag(3)
S[1, 1] <- 225
S[2, 2] <- 166.5
S[3, 3] <- 166.5
mu <- c(100, 100, 100)
# single random data set------------------------------------------------------
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu
)
str(X, list.len = 6)
# multiple random data sets---------------------------------------------------
X <- mvnram(
  n = 100,
  A = A,
  S = S,
  F = F,
  I = I,
  mu = mu,
  R = 10
)
str(X, list.len = 6)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
\code{\link[jeksterslabRsem:ram_Sigmatheta]{jeksterslabRsem::ram_Sigmatheta()}}, and \code{\link[jeksterslabRsem:ram_mutheta]{jeksterslabRsem::ram_mutheta()}}
for more information on the Reticular Action Model (RAM)
and \code{\link[=mvn]{mvn()}} for multivariate normal data generation.

Other multivariate data functions: 
\code{\link{mvnramsigma2}()},
\code{\link{mvn}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{multivariate data functions}
\keyword{multivariate}
\keyword{normal}
